moving ghillies into own file, removing old uniforms
