+++ addons/equipment/configs/uniforms/weapons/ghillies.hpp
@@ -1,0 +1,39 @@
+class U_B_FullGhillie_lsh;
+class SOCOMD_Ghillie_lsh : U_B_FullGhillie_lsh {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+class SOCOMD_AMCU_Ghillie_lsh : U_B_FullGhillie_lsh {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+
+class U_B_FullGhillie_sard;
+class SOCOMD_Ghillie_sard : U_B_FullGhillie_sard {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+class SOCOMD_AMCU_Ghillie_sard : U_B_FullGhillie_sard {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+
+class U_B_FullGhillie_ard;
+class SOCOMD_Ghillie_ard : U_B_FullGhillie_ard {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+class SOCOMD_AMCU_Ghillie_ard : U_B_FullGhillie_ard {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+
+class U_B_T_FullGhillie_tna_F;
+class SOCOMD_Ghillie_jngl : U_B_T_FullGhillie_tna_F {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
+class SOCOMD_AMCU_Ghillie_jngl : U_B_T_FullGhillie_tna_F {
+    scope=1;
+    SOCOMD_Ghillie = 1;
+};
